se agrega sombra al navbar
